--- front/src/App.css
@@ -1,30 +1,0 @@
-.gradient-custom-2 {
-  /* fallback for old browsers */
-  background: #4b79a1;
-
-  /* Chrome 10-25, Safari 5.1-6 */
-  background: -webkit-linear-gradient(to right, #4b79a1, #283e51, #1e3c72, #2a5298, #1c92d2);
-
-  /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
-  background: linear-gradient(to right, #4b79a1, #283e51, #1e3c72, #2a5298, #1c92d2);
-}
-
-@media (min-width: 768px) {
-  .gradient-form {
-    height: 100vh !important;
-  }
-}
-
-@media (min-width: 769px) {
-  .gradient-custom-2 {
-    border-top-right-radius: .3rem;
-    border-bottom-right-radius: .3rem;
-  }
-}
-
-.full-height-container {
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
